requirements.txt
alembic==1.0.5
APScheduler==3.5.3
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.3.5
Cerberus==1.2
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
Click==7.0
cookies==2.2.1
cryptography==2.3.1
decorator==4.3.0
Django==2.1.3
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
flask-jsondash==6.3.3
Flask-JWT==0.3.2
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
googlemaps==3.0.2
idna==2.7
imgurpython==1.1.7
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
nbformat==4.4.0
numpy==1.15.3
pandas==0.23.4
pipenv==2018.7.1
plotly==3.3.0
pluggy==0.7.1
py==1.6.0
pycparser==2.18
PyJWT==1.4.2
pylint==2.1.1
pymongo==3.7.2
PyMySQL==0.9.2
pytest==3.8.0
python-dateutil==2.7.5
python-editor==1.0.3
python-i18n==0.3.3
pytz==2018.5
PyYAML==3.13
requests>=2.20.0
responses==0.9.0
retrying==1.3.3
six==1.11.0
SQLAlchemy==1.2.14
toml==0.9.6
tox==3.3.0
traitlets==4.3.2
typed-ast==1.1.0
tzlocal==1.5.1
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
